% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{bundle}
\alias{bundle}
\title{Eric Edward Bryant's bundle of R packages}
\usage{
bundle()
}
\value{
Returns a named list with the following elements:

1. `imports`    -- vector of directly imported packages. These are my primary
                   system dependencies.
2. `remotes`    -- vector of "remote" packages that are not available on
                   CRAN. These are mostly my personal GitHub hosted packages.
3. `downstream` -- vector of packages that are required by the directly
                   imported packages specified in `imports` and `remotes`.
}
\description{
This package bundles my frequently used R packages to simplify
package reinstallation. The `bundle()` function lists the packages that are
installed as part of this bundle package.
}
