% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{install_remotes}
\alias{install_remotes}
\title{Install GitHub remotes}
\usage{
install_remotes(...)
}
\arguments{
\item{...}{Further arguments passed to \link[remotes:install_github]{remotes::install_github}}
}
\description{
I had to add this because the remotes are not being installed for some
reason.
}
